FIX Restrict Time Zone Alert Window
